Split e2e tests into smaller files by feature area
The single 1566-line e2e_test.go was becoming unwieldy. Split into 11
focused files: helpers, basic, control flow, concurrency, replicators,
arrays, procs/functions, types, protocols, records, and preprocessor.
